refactor: split test_fleet into utils for sample functions, test_models and test_views
