+++ fleet/tests/test_views.py
@@ -11,31 +11,11 @@ from fleet.serializers import (
     AirplaneDetailSerializer,
     AirplaneSerializer,
 )
+from fleet.tests.utils import sample_airplane_type, sample_airplane
+from user.tests.utils import sample_user
 
 AIRPLANE_TYPE_URL = reverse("fleet:airplane-type-list")
 AIRPLANE_URL = reverse("fleet:airplane-list")
-
-
-def sample_airplane_type(**params):
-    defaults = {
-        "name": "Boeing 737-800",
-    }
-    defaults.update(params)
-
-    return AirplaneType.objects.create(**defaults)
-
-
-def sample_airplane(**params):
-    """Create a sample airplane with its type passed as a parameter."""
-
-    defaults = {
-        "name": "EI-EPC",
-        "rows": 30,
-        "seats_in_row": 6,
-    }
-    defaults.update(params)
-
-    return Airplane.objects.create(**defaults)
 
 
 def airplane_type_detail_url(airplane_type_id):
@@ -44,26 +24,6 @@ def airplane_type_detail_url(airplane_type_id):
 
 def airplane_detail_url(airplane_id):
     return reverse("fleet:airplane-detail", args=[airplane_id])
-
-
-class AirplaneTypeModelTest(TestCase):
-    def test_airplane_type_str(self):
-        airplane_type = sample_airplane_type()
-        self.assertEqual(str(airplane_type), airplane_type.name)
-
-
-class AirplaneModelTest(TestCase):
-    def setUp(self):
-        self.airplane_type = sample_airplane_type()
-        self.airplane = sample_airplane(airplane_type=self.airplane_type)
-
-    def test_plane_capacity(self):
-        capacity = self.airplane.rows * self.airplane.seats_in_row
-        self.assertEqual(self.airplane.plane_capacity, capacity)
-
-    def test_airplane_str(self):
-        expected_str = f"{self.airplane.name}, type: {self.airplane_type.name}"
-        self.assertEqual(str(self.airplane), expected_str)
 
 
 class PublicAirplaneTypesAPITests(TestCase):
@@ -145,9 +105,7 @@ class PublicAirplaneAPITests(TestCase):
         self.assertEqual(response.data, serializer.data)
 
     def test_create_airplane_forbidden(self):
-        user = get_user_model().objects.create_user(
-            email="test@test.com", password="test1234"
-        )
+        user = sample_user()
         self.client.force_authenticate(user)
 
         payload = {
